+++ src/lib/Adaboost.cpp
@@ -15,7 +15,7 @@ WeakClassifier::WeakClassifier(BacktrackingAlgorithm<WeightedError, double> &A)
 
 // ===== Adaboost
 
-Adaboost::Adaboost(WeightedDataset<int> &d, DTOptions &opt)
+Adaboost::Adaboost(WeightedDataset<double> &d, DTOptions &opt)
     : options(opt), dataset(d), algo(opt) {
   algo.options.verbosity = DTOptions::SILENT;
 }
@@ -99,8 +99,6 @@ void Adaboost::initialize_weights() {
     for (auto x : X) {
       weight[y].push_back(double(X.weight(x)));
       algo.addExample(X[x], y, weight[y].back());
-      //reminder
-      //printf("%f ", weight[y][x]);
     }
   }
 }
@@ -119,15 +117,11 @@ void Adaboost::update_weights() {
       double upred = alpha * (last_tree.predict(X[xi]) ? 1 : -1);
       weight[y][i] *= (exp(-u * upred)/ (2 * sqrt(err * (1 - err)))); // I'm not sure what this is yet
       algo.setWeight(y, i, weight[y][i]);
-      //reminder
-      //printf("%f ", weight[y][i]);
-
       ++i;
     }
   }
 }
 
-// reminder
 std::vector<WeakClassifier> Adaboost::getClassifier() {
   return classifiers;
 }
